pyproject.toml
[project]
name = "chronicle"
version = "0.1.0"
description = "A perspective engine - historical context for present decisions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    "numpy>=1.24.0",
    "sqlite-vec>=0.1.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
chronicle = "chronicle.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
